requirements.txt
a2wsgi==1.10.8
agate==1.9.1
aiologic==0.14.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.0
apache-airflow-core==3.0.0
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-standard==1.1.0
apache-airflow-task-sdk==1.0.0
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
backports.strenum==1.3.1
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
build==1.2.2.post1
cadwyn==5.3.3
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.3
cycler==0.12.1
cymem==2.0.11
daff==1.3.46
dbt-adapters==1.14.8
dbt-common==1.23.0
dbt-core==1.9.4
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
debugpy==1.8.14
decorator==5.2.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
Flask==2.2.5
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.6
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.11.0
leather==0.4.0
libcst==1.7.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.3
more-itertools==10.7.0
msgpack==1.1.0
msgspec==0.19.0
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk
notebook==7.4.1
notebook_shim==0.2.4
numpy==2.2.5
oauthlib==3.2.2
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
ordered-set==4.1.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pendulum==3.1.0
pickleshare==0.7.5
pillow==11.2.1
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-argparse==1.7.0
rich-toolkit==0.14.4
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
snowplow-tracker==1.1.0
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
structlog==25.3.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tweepy
typer==0.15.3
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uri-template==1.3.0
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
wasabi==1.1.3
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==2.2.3
widgetsnbextension==4.0.14
wirerope==1.0.0
wrapt==1.17.2
yarg==0.1.9
zipp==3.21.0
praw
sqlalchemy
